Factor out a DbIsolationMixin from TestMetaData
I want to use the same behavior in another test class.
