+++ tests/base.py
@@ -157,6 +157,30 @@ class DbTestCase(unittest.TestCase):
         settings.BCRYPT_LOG_ROUNDS = cls._original_bcrypt_log_rounds
 
 
+class DbIsolationMixin(object):
+    """Use this mixin when test-level database isolation is desired.
+
+    For whatever reason (efficiency?), DbTestCase only wipes the database
+    during *class* setup and teardown. This leaks database state across test
+    cases, which smells pretty bad. Place this mixin before DbTestCase (or
+    derivatives, such as OsfTestCase) in your test class definition to wipe the
+    database during *test* setup and teardown.
+
+    """
+
+    def setUp(self):
+        super(DbIsolationMixin, self).setUp()
+        set_up_storage(
+            website.models.MODELS,
+            storage.MongoStorage,
+            addons=settings.ADDONS_AVAILABLE,
+        )
+
+    def tearDown(self):
+        super(DbIsolationMixin, self).tearDown()
+        teardown_database(database=database_proxy._get_current_object())
+
+
 class AppTestCase(unittest.TestCase):
     """Base `TestCase` for OSF tests that require the WSGI app (but no database).
     """
